Rename trakt module to shows
