+++ settings.gradle.kts
@@ -39,31 +39,31 @@ include(
     ":android:features:profile",
     ":shared:shared",
     ":shared:core:util",
+    ":shared:data:category:api",
+    ":shared:data:category:implementation",
     ":shared:data:database",
     ":shared:data:network",
     ":shared:data:datastore:api",
     ":shared:data:datastore:implementation",
     ":shared:data:datastore:testing",
-    ":shared:data:season-details:testing",
     ":shared:data:episodes:api",
     ":shared:data:episodes:implementation",
     ":shared:data:episodes:testing",
+    ":shared:data:season-details:api",
+    ":shared:data:season-details:implementation",
+    ":shared:data:season-details:testing",
     ":shared:data:similar:api",
     ":shared:data:similar:implementation",
     ":shared:data:similar:testing",
-    ":shared:data:season-details:api",
-    ":shared:data:season-details:implementation",
-    ":shared:data:category:api",
-    ":shared:data:category:implementation",
+    ":shared:data:shows:api",
+    ":shared:data:shows:implementation",
+    ":shared:data:shows:testing",
     ":shared:data:trailers:api",
     ":shared:data:trailers:implementation",
     ":shared:data:trailers:testing",
     ":shared:data:tmdb:api",
     ":shared:data:tmdb:implementation",
     ":shared:data:tmdb:testing",
-    ":shared:data:trakt:api",
-    ":shared:data:trakt:implementation",
-    ":shared:data:trakt:testing",
     ":shared:data:trakt-profile:api",
     ":shared:data:trakt-profile:implementation",
     ":shared:data:trakt-profile:testing",
