+++ shared/data/shows/api/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    id("tvmaniac.kmm.api")
+}
+
+kotlin {
+    android()
+    ios()
+
+    sourceSets {
+        sourceSets["commonMain"].dependencies {
+        }
+    }
+}
+
+android {
+    namespace = "com.thomaskioko.tvmaniac.shows.api"
+}
