+++ shared/data/shows/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/shows/implementation/di/TraktModule.kt
@@ -1,0 +1,5 @@
+package com.thomaskioko.tvmaniac.shows.implementation.di
+
+import org.koin.core.module.Module
+
+expect fun traktModule(): Module
