+++ shared/data/shows/testing/build.gradle.kts
@@ -1,0 +1,21 @@
+plugins {
+    id("tvmaniac.kmm.library")
+}
+
+kotlin {
+    android()
+    ios()
+
+    sourceSets {
+        sourceSets["commonMain"].dependencies {
+            implementation(project(":shared:core:util"))
+            implementation(project(":shared:data:shows:api"))
+            implementation(project(":shared:data:database"))
+            implementation(libs.coroutines.core)
+        }
+    }
+}
+
+android {
+    namespace = "com.thomaskioko.tvmaniac.shows.testing"
+}
