+++ shared/data/shows/testing/src/commonMain/kotlin/com/thomaskioko/tvmaniac/shows/testing/MockData.kt
@@ -1,0 +1,32 @@
+package com.thomaskioko.tvmaniac.shows.testing
+
+import com.thomaskioko.tvmaniac.core.db.SelectFollowedShows
+
+val cachedShowResult = listOf(
+    SelectFollowedShows(
+        trakt_id = 84958,
+        tmdb_id = 849583,
+        title = "Loki",
+        overview = "After stealing the Tesseract during the events of “Avengers: Endgame,” " +
+                "an alternate version of Loki is brought to the mysterious Time Variance " +
+                "Authority, a bureaucratic organization that exists outside of time and " +
+                "space and monitors the timeline. They give Loki a choice: face being " +
+                "erased from existence due to being a “time variant”or help fix " +
+                "the timeline and stop a greater threat.",
+        language = "en",
+        votes = 4958,
+        rating = 8.1,
+        genres = listOf("Horror", "Action"),
+        status = "Returning Series",
+        year = "2024",
+        runtime = 45,
+        poster_url = "/kEl2t3OhXc3Zb9FBh1AuYzRTgZp.jpg",
+        backdrop_url = "/kEl2t3OhXc3Zb9FBh1AuYzRTgZp.jpg",
+        aired_episodes = 12,
+        id = 84958,
+        synced = true,
+        created_at = 12345645,
+        trakt_id_ = 1232,
+        tmdb_id_ = 849583
+    )
+)
