--- shared/data/trakt/api/build.gradle.kts
@@ -1,17 +1,0 @@
-plugins {
-    id("tvmaniac.kmm.api")
-}
-
-kotlin {
-    android()
-    ios()
-
-    sourceSets {
-        sourceSets["commonMain"].dependencies {
-        }
-    }
-}
-
-android {
-    namespace = "com.thomaskioko.tvmaniac.trakt.api"
-}
