--- shared/data/trakt/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/implementation/di/TraktModule.kt
@@ -1,5 +1,0 @@
-package com.thomaskioko.tvmaniac.trakt.implementation.di
-
-import org.koin.core.module.Module
-
-expect fun traktModule(): Module
