+++ shared/shared/build.gradle.kts
@@ -37,8 +37,8 @@ kotlin {
             export(project(":shared:data:season-details:api"))
             export(project(":shared:data:category:api"))
             export(project(":shared:data:trailers:api"))
+            export(project(":shared:data:shows:api"))
             export(project(":shared:data:tmdb:api"))
-            export(project(":shared:data:trakt:api"))
             export(project(":shared:data:trakt-profile:api"))
             export(project(":shared:domain:following"))
             export(project(":shared:domain:discover"))
@@ -57,18 +57,18 @@ kotlin {
         sourceSets["commonMain"].dependencies {
 
             api(project(":shared:core:util"))
+            api(project(":shared:data:category:api"))
             api(project(":shared:data:database"))
             api(project(":shared:data:datastore:api"))
-            api(project(":shared:data:network"))
             api(project(":shared:data:episodes:api"))
+            api(project(":shared:data:network"))
             api(project(":shared:data:similar:api"))
             api(project(":shared:data:season-details:api"))
-            api(project(":shared:data:category:api"))
+            api(project(":shared:data:shows:api"))
             api(project(":shared:data:trailers:api"))
             api(project(":shared:data:tmdb:api"))
-            api(project(":shared:data:trakt:api"))
-            api(project( ":shared:data:trakt-profile:api"))
-            api(project( ":shared:data:trakt-service:api"))
+            api(project(":shared:data:trakt-profile:api"))
+            api(project(":shared:data:trakt-service:api"))
             api(project(":shared:domain:discover"))
             api(project(":shared:domain:following"))
             api(project(":shared:domain:seasondetails"))
@@ -76,16 +76,16 @@ kotlin {
             api(project(":shared:domain:show-details"))
             api(project(":shared:domain:trailers"))
 
+            implementation(project(":shared:data:category:implementation"))
             implementation(project(":shared:data:datastore:implementation"))
             implementation(project(":shared:data:episodes:implementation"))
             implementation(project(":shared:data:similar:implementation"))
             implementation(project(":shared:data:season-details:implementation"))
-            implementation(project(":shared:data:trailers:implementation"))
+            implementation(project(":shared:data:shows:implementation"))
             implementation(project(":shared:data:tmdb:implementation"))
-            implementation(project(":shared:data:trakt:implementation"))
+            implementation(project(":shared:data:trailers:implementation"))
             implementation(project(":shared:data:trakt-profile:implementation"))
             implementation(project(":shared:data:trakt-service:implementation"))
-            implementation(project(":shared:data:category:implementation"))
 
             implementation(libs.koin)
             implementation(libs.coroutines.core)
